+++ src/utils/systemReport.ts
@@ -57,6 +57,13 @@ const generalInfoReport = async () => {
         getDeviceLibContext()
     );
 
+    let jlinkVersion;
+    try {
+        jlinkVersion = getModuleVersion('JlinkARM', versions);
+    } catch {
+        jlinkVersion = getModuleVersion('jlink', versions);
+    }
+
     return [
         `- System:     ${manufacturer} ${model}`,
         `- BIOS:       ${vendor} ${version}`,
@@ -84,9 +91,7 @@ const generalInfoReport = async () => {
         `    - nrfjprog DLL: ${describeVersion(
             getModuleVersion('jprog', versions)
         )}`,
-        `    - JLink: ${describeVersion(
-            getModuleVersion('JlinkARM', versions)
-        )}`,
+        `    - JLink: ${describeVersion(jlinkVersion)}`,
         '',
     ];
 };